requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
neo4j>=5.15.0
sentence-transformers>=2.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
uagents>=0.4.0
scikit-learn>=1.3.0
